Refactoring and INIT
