+++ src/main/java/com/nefarious/socialnetwork/auth/dto/SessionResponse.java
@@ -1,0 +1,11 @@
+package com.nefarious.socialnetwork.auth.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class SessionResponse {
+    private String accessToken;
+    private String refreshToken;
+}
