+++ src/main/java/com/nefarious/socialnetwork/auth/dto/SigninRequest.java
@@ -1,0 +1,16 @@
+package com.nefarious.socialnetwork.auth.dto;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Pattern;
+import jakarta.validation.constraints.Size;
+import lombok.Data;
+
+@Data
+public class SigninRequest {
+    @Email @NotBlank private String email;
+    @NotBlank @Size(min = 8, max = 24)
+    @Pattern(regexp = "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,24}$",
+            message = "Password must contain at least one uppercase letter, one lowercase letter, one number, and one special character.")
+    private String password;
+}
