+++ src/main/java/com/nefarious/socialnetwork/auth/repository/OtpRepository.java
@@ -1,0 +1,7 @@
+package com.nefarious.socialnetwork.auth.repository;
+
+import com.nefarious.socialnetwork.auth.entity.OtpEntry;
+import org.springframework.data.repository.CrudRepository;
+
+public interface OtpRepository extends CrudRepository<OtpEntry, String> {
+}
