+++ src/main/java/com/nefarious/socialnetwork/auth/util/AuthEndpoint.java
@@ -1,0 +1,10 @@
+package com.nefarious.socialnetwork.auth.util;
+
+public class AuthEndpoint {
+    public static final String AUTH             = "/auth";
+    public static final String AUTH_SIGNUP      = "/signup";
+    public static final String AUTH_SIGNIN      = "/signin";
+    public static final String AUTH_VERIFY      = "/verify";
+    public static final String RESEND_OTP       = "/resend-otp";
+    public static final String REFRESH          = "/refresh";
+}
