+++ src/main/java/com/nefarious/socialnetwork/auth/util/enums/TokenType.java
@@ -1,0 +1,18 @@
+package com.nefarious.socialnetwork.auth.util.enums;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+/**
+ * Types of Tokens
+ * <br>
+ * ACCESS
+ * <br>
+ * REFRESH
+ */
+@AllArgsConstructor
+@Getter
+public enum TokenType {
+    ACCESS("access"), REFRESH("refresh");
+    private final String value;
+}
