--- src/main/java/com/nefarious/socialnetwork/dto/SessionResponse.java
@@ -1,10 +1,0 @@
-package com.nefarious.socialnetwork.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class SessionResponse {
-    private String accessToken;
-}
