+++ src/main/java/com/nefarious/socialnetwork/exceptions/BusinessException.java
@@ -1,0 +1,13 @@
+package com.nefarious.socialnetwork.exceptions;
+
+import com.nefarious.socialnetwork.util.interfaces.ErrorCode;
+import lombok.Getter;
+
+@Getter
+public class BusinessException extends RuntimeException {
+    private final ErrorCode errorCode;
+    public BusinessException(ErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+    }
+}
