--- src/main/java/com/nefarious/socialnetwork/exceptions/EmailSendException.java
@@ -1,10 +1,0 @@
-package com.nefarious.socialnetwork.exceptions;
-
-/**
- * Thrown when an e-mail cannot be sent.
- */
-public class EmailSendException extends RuntimeException {
-    public EmailSendException(String msg, Throwable cause) {
-        super(msg, cause);
-    }
-}
