--- src/main/java/com/nefarious/socialnetwork/repository/OtpRepository.java
@@ -1,7 +1,0 @@
-package com.nefarious.socialnetwork.repository;
-
-import com.nefarious.socialnetwork.entity.OtpEntry;
-import org.springframework.data.repository.CrudRepository;
-
-public interface OtpRepository extends CrudRepository<OtpEntry, String> {
-}
