--- src/main/java/com/nefarious/socialnetwork/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.nefarious.socialnetwork.repository;
-import com.nefarious.socialnetwork.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-import java.util.UUID;
-
-public interface UserRepository extends JpaRepository<User, UUID> {
-    Optional<User> findByEmail(String email);
-    Optional<User> findByUsername(String username);
-}
