+++ src/main/java/com/nefarious/socialnetwork/user/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package com.nefarious.socialnetwork.user.repository;
+import com.nefarious.socialnetwork.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+import java.util.UUID;
+
+public interface UserRepository extends JpaRepository<User, UUID> {
+    Optional<User> findByEmail(String email);
+    Optional<User> findByUsername(String username);
+}
