--- src/main/java/com/nefarious/socialnetwork/util/Constants.java
@@ -1,5 +1,0 @@
-package com.nefarious.socialnetwork.util;
-
-public class Constants {
-
-}
