--- src/main/java/com/nefarious/socialnetwork/util/Endpoints.java
@@ -1,7 +1,0 @@
-package com.nefarious.socialnetwork.util;
-
-public class Endpoints {
-    public static final String AUTH = "/auth";
-    public static final String AUTH_SIGNUP = "/signup";
-    public static final String AUTH_VERIFY = "/verify";
-}
