+++ src/main/java/com/nefarious/socialnetwork/util/interfaces/ErrorCode.java
@@ -1,0 +1,8 @@
+package com.nefarious.socialnetwork.util.interfaces;
+
+import org.springframework.http.HttpStatus;
+
+public interface ErrorCode {
+    String getMessage();
+    HttpStatus getHttpStatus();
+}
