+++ src/main/resources/application.properties
@@ -42,7 +42,9 @@ spring.mail.properties.mail.smtp.ssl.enable=false
 # ----------------- JWT Setting --------------------------------------#
 jwt.secret=${JWT_SECRET}
 # 1 hr in ms
-jwt.expiration-ms=3600000
+jwt.access-expiration-ms=3600000
+# 6 months in ms
+jwt.refresh-expiration-ms=15778476000
 
 
 # ----------------- Dev Only --------------------------------------#
